+++ webgrid/tests/test_columns.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 import datetime as dt
 from decimal import Decimal as D
 from blazeutils.containers import HTMLAttributes