+++ webgrid/webgrid_nose.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 import os
 from os import path as osp