+++ webgrid_ta/model/__init__.py
@@ -1,7 +1,9 @@
+from __future__ import absolute_import
 import datetime as dt
 from decimal import Decimal as D
 
 from flask.ext.sqlalchemy import SQLAlchemy
+from six.moves import range
 
 db = SQLAlchemy()